requirements.txt
numpy==1.16.4
nltk==3.4.4
PyYAML==5.3
pandas==0.25.3
scikit-learn==0.22.1
textblob==0.15.3
matplotlib==3.1.2
torch==1.3.1
